make sure we finalize any internal na+sm mercury we create
calling nexus_bootstrap() with a NULL localhandle causes us to
create our own internal instance of a na+sm mercury.  we need
to shutdown anything we create, so we need to track this instance
and finalize it at nexus_destroy() time.  note that you cannot
extend the life of an internal na+sm by duping the progressor
returned by nexus_localprogressor().